+++ compose-app/src/main/java/com/mapbox/maps/compose/testapp/examples/annotation/DynamicViewAnnotationActivity.kt
@@ -7,6 +7,7 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -419,7 +420,11 @@ public class DynamicViewAnnotationActivity : ComponentActivity() {
       modifier = Modifier
         .width(120.dp)
         .height(60.dp)
-        .clickable(onClick = onClick)
+        .clickable(
+          interactionSource = remember { MutableInteractionSource() },
+          indication = null,
+          onClick = onClick
+        )
     ) {
       Image(
         modifier = Modifier
@@ -453,7 +458,11 @@ public class DynamicViewAnnotationActivity : ComponentActivity() {
       modifier = Modifier
         .width(180.dp)
         .height(70.dp)
-        .clickable(onClick = onClick)
+        .clickable(
+          interactionSource = remember { MutableInteractionSource() },
+          indication = null,
+          onClick = onClick
+        )
     ) {
       Image(
         modifier = Modifier